refactor : meetup_board type 위치 변경 및 모임생성모달 ui 개선 #150
