+++ src/app/datepicker.scss
@@ -5,7 +5,7 @@
 }
 
 .react-datetime-picker__wrapper {
-  padding: 0.7rem 2rem;
+  padding: 0.7rem 1.2rem;
   border-radius: 18px;
   border: none;
 }
